@gsouvik Feature devsupport merge complete

+++ Web/js/app/controllers/technician_manager.js
@@ -8,6 +8,31 @@
  */
 $(document).ready(function() {
   $(".btn-warning").hide();
+  $("#document-upload").hide();
+  $("#documents").hide();
+  Dropzone.autoDiscover = false;
+  if($("#document-upload").length>0){
+    var dropzone = new Dropzone("#document-upload");
+    dropzone.on("success", function(event,res) {
+      if(res.result == 0) {
+        toastr.error(res.message);
+        dropzone.removeAllFiles();
+      } else {
+        toastr.success(res.message);
+        $("#document-upload").hide();
+        dropzone.removeAllFiles();
+        technician_manager.loadPhoto('technician_id',parseInt( $("input[name=\"itemId\"]").val()));
+      }
+    });
+  }
+  $("#tech_photo_upload").on('click',function(e){
+    e.preventDefault();
+    $("#document-upload").show();
+  });
+  $(document).on('click','.remove-image',function(e){
+    e.preventDefault();
+    technician_manager.removePhoto($(this).data("id"));
+  });
   $.contextMenu({
     selector: '.select_item',
     callback: function(key, options) {
@@ -103,19 +128,19 @@ $(document).ready(function() {
   });//Edit a technician
 
   $("#btn_delete_technician").click(function() {
-		var msg = $('#confirmmsg-delete').val();
-		if (typeof msg !== typeof undefined && msg !== false) {
-			utils.showConfirmBox(msg, function(result){
-				if(result)
-				{
-					technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
-				}
-			});
-		}
-		else
+	var msg = $('#confirmmsg-delete').val();
+	if (typeof msg !== typeof undefined && msg !== false) {
+	  utils.showConfirmBox(msg, function(result){
+		if(result)
 		{
 			technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
 		}
+	  });
+	}
+	else
+	{
+	  technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
+	}
   });//Delete a technician
 
   if (utils.getQueryVariable("mode") === "edit") {
@@ -123,6 +148,9 @@ $(document).ready(function() {
     $(".welcome").fadeOut('2000').removeClass("show").addClass("hide");
     $(".technician_add").hide();
     technician_manager.getItem(utils.getQueryVariable("technician_id"));
+  } else {
+    $("#document-upload").hide();
+    $("#tech_photo_upload").parent("li").hide();
   }//Load technician
 
 //  if (utils.getQueryVariable("mode") === "add" && utils.getQueryVariable("test") === "true") {
@@ -202,7 +230,10 @@ $(document).ready(function() {
     $("input[name=\"technician_name\"]").val(dataWs.technician.technician_name);
     $("input[name=\"technician_phone\"]").val(dataWs.technician.technician_phone);
     $("input[name=\"technician_email\"]").val(dataWs.technician.technician_email);
-//    $("input[name=\"technician_document\"]").val(dataWs.technician.technician_document);
+    $("input[name=\"itemCategory\"]").val('technician_id');
+    $("input[name=\"itemId\"]").val(parseInt(dataWs.technician.technician_id));
+    $("input[name=\"itemReplace\"]").val(true);
+    technician_manager.loadPhoto('technician_id',parseInt(dataWs.technician.technician_id));
     $("input[name=\"technician_active\"]").prop('checked', utils.setCheckBoxValue(dataWs.technician.technician_active));
   };
   technician_manager.delete = function(technician_id) {
@@ -253,5 +284,40 @@ $(document).ready(function() {
       }
     });
   };
+
+  technician_manager.loadPhoto = function(itemCategory, itemId) {
+    datacx.post("file/load", {"itemCategory": itemCategory, "itemId": itemId}).then(function(reply){
+      if (reply === null || reply.result === 0) {//has an error
+        toastr.error(reply.message);
+        return undefined;
+      } else {//success
+        toastr.success(reply.message);
+        if(reply.fileResults.length>0){
+          $.each(reply.fileResults, function(key, file){
+            var lightboxImage = utils.createImageLightboxElement(file.webPath, itemId, file.title);
+            var remove = utils.createRemoveFileElement(file.document_id);
+            $("#documents").append('<div class="document-block" id="document-'+file.document_id+'">'+lightboxImage+remove+'</div>');
+            $("#documents").show();
+          });
+        } else {
+          $("#document-upload").show();
+        }
+      }
+    });
+  }
+
+  technician_manager.removePhoto = function(document_id) {
+    datacx.post("file/remove", {"document_id": document_id, "itemCategory": 'technician_id'}).then(function(reply){
+      if (reply === null || reply.result === 0) {//has an error
+        toastr.error(reply.message);
+        return undefined;
+      } else {//success
+        toastr.success(reply.message);
+        $("#documents").hide();
+        $("#document-"+document_id).remove();
+        $("#document-upload").show();
+      }
+    });
+  }
 
 }(window.technician_manager = window.technician_manager || {}));
